aeneas/functions/aeneas/requirements.txt
paho-mqtt
aeneas
requests

